+++ app/f1/party/[partyId]/page.tsx
@@ -1,0 +1,14 @@
+import { PartyPageClient } from "../../../PartyPageClient";
+
+export const dynamic = "force-dynamic";
+
+type PartyPageProps = {
+  params: Promise<{
+    partyId: string;
+  }>;
+};
+
+export default async function PartyPage({ params }: PartyPageProps) {
+  const { partyId } = await params;
+  return <PartyPageClient partyId={partyId} />;
+}